Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -15,7 +15,7 @@ int main()
         printf("%d",a*b);
     }
     else if (o=='/'){
-        printf("%.2f",a/b);
+        printf("%f",(float)a/b);
     }
     return 0;
 }
